requirements.txt
GitPython
openai
python-dotenv
qdrant-client
requests
esprima